Move token endpoints to web ui from director
