+++ director/director.go
@@ -41,7 +41,6 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/pelicanplatform/pelican/config"
-	"github.com/pelicanplatform/pelican/database"
 	"github.com/pelicanplatform/pelican/metrics"
 	"github.com/pelicanplatform/pelican/param"
 	"github.com/pelicanplatform/pelican/pelican_url"
@@ -81,11 +80,6 @@ type (
 
 	// Context key for the project name
 	ProjectContextKey struct{}
-
-	CreateApiTokenReq struct {
-		Name      string `json:"name"`
-		CreatedBy string `json:"created_by"`
-	}
 )
 
 const (
@@ -1416,65 +1410,6 @@ func listNamespacesV2(ctx *gin.Context) {
 	ctx.JSON(http.StatusOK, namespacesAdsV2)
 }
 
-func createApiToken(ctx *gin.Context) {
-	authOption := token.AuthOption{
-		Sources: []token.TokenSource{token.Cookie},
-		Issuers: []token.TokenIssuer{token.LocalIssuer},
-		Scopes:  []token_scopes.TokenScope{token_scopes.WebUi_Access},
-	}
-
-	status, ok, err := token.Verify(ctx, authOption)
-	if !ok {
-		log.Warningf("Cannot verify token: %v", err)
-		ctx.JSON(status, server_structs.SimpleApiResp{
-			Status: server_structs.RespFailed,
-			Msg:    err.Error(),
-		})
-		return
-	}
-	// marshall body into struct
-	var req CreateApiTokenReq
-	err = ctx.ShouldBindJSON(&req)
-	if err != nil {
-		ctx.JSON(http.StatusBadRequest, server_structs.SimpleApiResp{
-			Status: server_structs.RespFailed,
-			Msg:    fmt.Sprintf("Invalid request body: %v", err),
-		})
-		return
-	}
-
-	scopes := fmt.Sprintf("%s,%s", token_scopes.Monitoring_Query.String(), token_scopes.Monitoring_Scrape.String())
-	token, err := database.CreateApiKey(req.Name, req.CreatedBy, scopes)
-	if err != nil {
-		log.Warning("Failed to create API key: ", err)
-		ctx.JSON(status, server_structs.SimpleApiResp{
-			Status: server_structs.RespFailed,
-			Msg:    err.Error(),
-		})
-		return
-	}
-
-	ctx.JSON(http.StatusOK, gin.H{"token": token})
-}
-
-func deleteApiToken(ctx *gin.Context) {
-	id := ctx.Param("id")
-	err := database.DeleteApiKey(id)
-	if err != nil {
-		log.Warning("Failed to delete API key: ", err)
-		ctx.JSON(http.StatusInternalServerError, server_structs.SimpleApiResp{
-			Status: server_structs.RespFailed,
-			Msg:    err.Error(),
-		})
-		return
-	}
-
-	ctx.JSON(http.StatusOK, server_structs.SimpleApiResp{
-		Status: server_structs.RespOK,
-		Msg:    "API key deleted",
-	})
-}
-
 func getPrefixByPath(ctx *gin.Context) {
 	pathParam := ctx.Param("path")
 	if pathParam == "" || pathParam == "/" {
@@ -1694,8 +1629,6 @@ func RegisterDirectorAPI(ctx context.Context, router *gin.RouterGroup) {
 		// Rename the endpoint to reflect such plan.
 		directorAPIV1.GET("/discoverServers", discoverOriginCache)
 
-		directorAPIV1.POST("/createApiToken", createApiToken)
-		directorAPIV1.DELETE("/deleteApiToken/:id", deleteApiToken)
 	}
 
 	directorAPIV2 := router.Group("/api/v2.0/director", web_ui.ServerHeaderMiddleware)
